[Correct] LetterView -> SendLetterView 명칭 변경
SendLetterView 이름 변경 및 올바른 폴더로 이동
